HW1 Refactoring 2
